sort imports and add precommit
