+++ gsw_xarray/tests/test_utils.py
@@ -1,5 +1,3 @@
-import pytest
-
 from .._function_utils import args_and_kwargs_to_kwargs
